refactor: share control panel events without three dependency
